chore: db.js 타입스크립트로 변경
